affinitas-backend/pyproject.toml
[project]
name = "affinitas-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.29.0",
    "fastapi[standard]>=0.115.12",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.7",
    "langsmith>=0.3.42",
    "motor>=3.7.1",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.13.0",
    "slowapi>=0.1.9",
]
